Switch to Arc<T> based system
Instead of maintaining everything in memory permanently (which was a hack fix), everything is stored in an Arc. This will hopefully maintain long term program stability.

As well as using Arc's a new system for how data is handled during runtime was introduced. Instead of maintaining all objects in memory only hashes are stored. At the same time all new data is put into a separate disk queue for saving.

This system also provides some sort of redundancy, in the event that a bitcode blob is corrupted not all data on a shard will be lost.

This is a very early wip implementation, but has promising performance and stability.